Update schemas and example to support discovery-options parameter
Upstream schemas have changed and `options` is now a required field in
the request. The options field can be used to pass arbitrary parameters
through to the external discovery method.

The static example has been updated to demonstrate how a `limit`
parameter can be optionally passed through to artificially limit the
number of nodes matched by the filters.

https://github.com/choria-io/schemas/pull/64